jotai magiaa

+++ src/modules/arithmetic.py
@@ -1,21 +1,21 @@
 def plus_handler(a, b):
     (a, b) = convert_arguments(a, b)
-    return a + b
+    return int(a + b)
 
 
 def minus_handler(a, b):
     (a, b) = convert_arguments(a, b)
-    return a - b
+    return int(a - b)
 
 
 def multiply_handler(a, b):
     (a, b) = convert_arguments(a, b)
-    return a * b
+    return int(a * b)
 
 
 def division_handler(a, b):
     (a, b) = convert_arguments(a, b)
-    return a / b
+    return int(a / b)
 
 
 # Convert arguments to integers
